+++ examples/workflows/cockroach_folder/main.tf
@@ -13,10 +13,10 @@ variable "cluster_name" {
   nullable = false
 }
 
-variable "serverless_spend_limit" {
+variable "provisioned_capacity" {
   type     = number
   nullable = false
-  default  = 0
+  default  = 1000
 }
 
 variable "cloud_provider" {
@@ -74,7 +74,9 @@ resource "cockroach_cluster" "example" {
   name           = var.cluster_name
   cloud_provider = var.cloud_provider
   serverless = {
-    spend_limit = var.serverless_spend_limit
+    usage_limits = {
+      provisioned_capacity = var.provisioned_capacity
+    }
   }
   regions   = [for r in var.cloud_provider_regions : { name = r }]
   parent_id = cockroach_folder.example_folder_parent.id